minor display edits, added cart summary page
edited number format for price and total displays
added cart summary page
fixed the admin nav bar
fixed the back button for cart summary and order summary

+++ db.sql
@@ -78,7 +78,7 @@ CREATE TABLE `product`(
 	`description` text NOT NULL,
 	`image` longblob NOT NULL,
 	`stock` int(15) NOT NULL,
-	`price` int(15) NOT NULL,
+	`price` float NOT NULL,
 	`productUnitID` int(15) NOT NULL,
 	`productCategory` int(15),
 	`seller` int(15) NOT NULL,
